+++ packages/php-laravel/src/SSOClientServiceProvider.php
@@ -9,20 +9,8 @@ class SSOClientServiceProvider extends ServiceProvider
 {
     public function boot()
     {
-        $this->publishControllers();
-        $this->publishConfig();
-    }
-
-    protected function publishControllers()
-    {
         $this->publishes([
             __DIR__ . '/SSOController.php' => App::path('Http/Controllers/SSO/SSOController.php'),
-        ], 'sso-config');
-    }
-
-    protected function publishConfig()
-    {
-        $this->publishes([
             __DIR__ . '/sso.php' => config_path('sso.php'),
         ], 'sso-config');
     }